relocate PagingResponseIterator
